// src/app/modules/hcm/shared/models/concepts/accounting-template-detail.ts
export class AccountingTemplateDetail {

    idAccountingTemplateDetail: number = -1;
    idAccountingItemDetail: number = -1;
    idAssigned: number = -1;
    idSegmentType: number = -1;
    idReplace: boolean = false;
    idOmit: boolean = false;
    assignedValue: string = "";
    identifier: string = "";
	ordinal: number = 0;
	description: string = "";
}